% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parser.R
\name{get_schema_version}
\alias{get_schema_version}
\title{get_schema_version}
\usage{
get_schema_version()
}
\value{
a character
}
\description{
Extract the release version of the MIIID schema used for validation.
}
