% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_example_miiid.R
\docType{data}
\name{invalid_example_miiid}
\alias{invalid_example_miiid}
\title{Invalid MIIID data example}
\format{
list
}
\source{
based on <https://github.com/FAIR-MI/miiid-schema/blob/790f69ddd6fa8a64232e519b73c68090c64e6041/src/data/examples/invalid/IntermicrobialInteraction-missing-id.yaml>
}
\usage{
invalid_example_miiid
}
\description{
One instance of IntermicrobialInteraction object that does not fits the schema
}
\keyword{datasets}
