% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_example_miiid.R
\docType{data}
\name{valid_example_miiid}
\alias{valid_example_miiid}
\title{Valid MIIID data example}
\format{
list
}
\source{
<https://github.com/FAIR-MI/miiid-schema/blob/790f69ddd6fa8a64232e519b73c68090c64e6041/src/data/examples/valid/IntermicrobialInteraction-pairwise.yaml>
}
\usage{
valid_example_miiid
}
\description{
One instance of IntermicrobialInteraction object that fits the schema
}
\keyword{datasets}
